wrap ConvolvedFluxPlugin to allow use in forced measurement
Change the ConvolvedFluxPlugin class into a base class that we then
wrap with the appropriate plugin base class. There's also a little
bit of fiddling required to give the forced measurement a slightly
different signature.

In order to do the forced measurement, we need to add transformation
from the reference frame to the measurement frame. Since we're
already doing forced measurement (center is provided, aperture shape
and radius are provided), moved the calculation into a 'measureForced'
method, and have 'measure' call that with a null transformation.

Modified the test to check that forced measurement also works.