fix undeblended measurement
Would get the following exception when activating convolved measurements
of undeblended sources:

lsst::pex::exceptions::InvalidParameterError: 'Field with name 'ext_convolved_ConvolvedFlux_0_deconv' already present in schema.'

This was due to assuming the algorithm name was set in stone.

Extended tests to catch this and test the undeblended measurements as well.